DashBoardView Study

+++ common_lib/src/main/java/pan/lib/common_lib/utils/ext/DimensionExt.kt
@@ -19,7 +19,7 @@ val Number.dp2px: Float
         TypedValue.COMPLEX_UNIT_DIP,
         this.toFloat(),
         Resources.getSystem().displayMetrics
-    ).toFloat()
+    )
 
 
 val Number.px2dp: Int
